% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{mx.crypto-package}
\alias{mx.crypto}
\alias{mx.crypto-package}
\title{\packageTitle{mx.crypto}}
\description{\packageDescription{mx.crypto}}
\section{Package Content}{\packageIndices{mx.crypto}}
\author{\packageAuthor{mx.crypto}}
\section{Maintainer}{\packageMaintainer{mx.crypto}}
