% tinyrox says don't edit this manually, but it can't stop you!
\name{mxc_megolm_outbound_pickle}
\alias{mxc_megolm_outbound_pickle}
\title{Pickle an outbound group session}
\usage{
mxc_megolm_outbound_pickle(gs, key)
}
\arguments{
\item{gs}{An outbound GroupSession.}

\item{key}{32-byte raw vector.}
}
\value{
Base64 string.
}
\description{
Pickle an outbound group session
}
