% tinyrox says don't edit this manually, but it can't stop you!
\name{mxc_signing_key_public}
\alias{mxc_signing_key_public}
\title{Read a signing key's public Ed25519 key}
\usage{
mxc_signing_key_public(key)
}
\arguments{
\item{key}{A signing-key handle.}
}
\value{
Unpadded base64 Ed25519 public key.
}
\description{
Read a signing key's public Ed25519 key
}
